VIP Member Access
created a members button next to logout button
this button will only be visible if user's are logged in

This is implemented using angular route gaurd.
which has common interface canActivate

we can have custom implementation or use OktaAuthGaurd
here i have used OktaAuthGuard which will give access to members page
to logged in users only. If we enter the url in browser without
logging in then OktaAuthGuard will redirect to login page